Fri, Jan 6, 2023, 7:43 PM +00:00

+++ sentimental-cash/cash.py
@@ -3,16 +3,16 @@ from cs50 import get_float
 
 #Define the funcctions
 def calculate_quarters(dollars):
-    return int(dollars * 4)
+    return dollars * 4
 
 def calculate_dimes(dollars):
-    return int(dollars * 10)
+    return dollars * 10
 
 def calculate_nickels(dollars):
-    return int(dollars * 20)
+    return dollars * 20
 
 def calculate_pennies(dollars):
-    return int(dollars * 100)
+    return dollars * 100
 
 #Ask how many dollars the customer is owed
 while True:
